Stream (#33)
* stream and handle of lbl

* unified query and s query

* added stream multi match

* type erasure

* rename interface classes

* update python interface with stream functions

* added comments and minor fixes

* update pypi actions

* python tests

* update py version

* javascript wip

* fix overflow and update python package

* fix findone exception

* improve tests and fix in match interface

* fixes in python compilation and multi check

* update pypi action and prod py version

* added check mediators

* delete datasets and remove allocations in tests

+++ src/rematch/evaluation/extended_va/dfa/extended_det_va_state.hpp
@@ -24,7 +24,7 @@ class ExtendedDetVAState {
   std::vector<std::unique_ptr<std::vector<CaptureSubsetPair>>>
       cached_transitions{256};
   ECSNode* output_node = nullptr;
-  int phase = -1;
+  int64_t phase = -1;
 
   ExtendedDetVAState();
 
@@ -40,7 +40,7 @@ class ExtendedDetVAState {
   void set_node(ECSNode* node);
   void unset_node();
   ECSNode* get_node() const;
-  void set_phase(int phase);
+  void set_phase(int64_t phase);
 
   inline size_t get_subset_size() {return states_subset_.size(); }
